+++ src/main/java/roger/Ui.java
@@ -1,3 +1,5 @@
+package roger;
+
 import java.util.Scanner;
 
 public class Ui {
