+++ src/main/java/roger/commands/ExitCommand.java
@@ -1,4 +1,9 @@
-import java.time.LocalDate;
+package roger.commands;
+
+import roger.Storage;
+import roger.TaskList;
+import roger.Ui;
+import roger.commands.Command;
 
 public class ExitCommand extends Command {
