+++ src/main/java/roger/commands/UnknownCommand.java
@@ -1,3 +1,9 @@
+package roger.commands;
+
+import roger.Storage;
+import roger.TaskList;
+import roger.Ui;
+import roger.commands.Command;
 
 public class UnknownCommand extends Command {
