+++ src/main/java/roger/tasks/Event.java
@@ -1,3 +1,5 @@
+package roger.tasks;
+
 import java.time.LocalDate;
